# Compare `any` arrays