We can try to fix https://github.com/expr-lang/expr/issues/371 with nested comparison of arrays' values but I'm not sure if it's worth it to increase amount of switch-cases for such rare (imo) case.

Also, I found that expr always goes to `reflect.DeepEqual` for bools and fixed it.