@bizywizy 2024-02-28
Okay, I opened separate PR for bools [here](https://github.com/expr-lang/expr/pull/585).

>But yes, I think we should not add cases for arrays.

We can add `equalArrays` just before `reflect.DeepEqual` and then we won't affect main switch-case statement
```golang
func Equal(a, b any) {
switch x := a.(type) {
    ...
}
if res, ok := equalArrays(a, b); ok {
    return res
}
return reflect.DeepEqual(a, b)
```
but decrease performance of non-primitive comparisons. Do you have any thoughts how frequently it happens?



>Maybe a builtin deepEqual is needed which will be ignoring types of arrays (and maps?).

Could you please give me couple examples? I think I didn't get it.