@antonmedv 2024-02-28
> We can add equalArrays just before reflect.DeepEqual and then we won't affect main switch-case statement

Sounds like a good solution. Make sure it is a deep equal as well: `[1,[2,[3]]]==[1,[2,[3]]]`


> Do you have any thoughts how frequently it happens?

I don't know ¯\\\_(ツ)_/¯ 
I will be cool actually to vigure out what **reflect.DeepEqual** covers and add tests for those cases to make sure we will not break those in the future. 

I also have an idea to test: add **OpEqualInt** & **OpEqualBool** & **OpEqualString** opcodes. I belive those will be the most common once. **BUT**, I'm not sure what adding more opcodes will be faster than **OpEqual** switch, so those changes should be benchmarked.