@bizywizy 2024-03-04
>I don't know ¯\_(ツ)_/¯
I will be cool actually to vigure out what reflect.DeepEqual covers and add tests for those cases to make sure we will not break those in the future.

The only case I found in my codebase involves `fmt.Stringer` objects. We can handle them at compile time (by calling .String() and then checking for equality) or simply add another switch-case to `runtime.Equal`.

>I also have an idea to test: add OpEqualInt & OpEqualBool & OpEqualString opcodes. I belive those will be the most common once. BUT, I'm not sure what adding more opcodes will be faster than OpEqual switch, so those changes should be benchmarked.

Expr already includes `OpEqualString` and `OpEqualInt` opcodes so I compared these with `OpEqual` which is handled by the switch-case in `runtime.Equal`.

`OpEqualString` + `OpEqualInt` slightly better than `OpEqual`
```
❯ benchstat master.txt only-op-equal.txt                                                                                                                                                                                     (gp-eu-central-1/delta)
goos: darwin
goarch: arm64
pkg: github.com/expr-lang/expr
                │ master.txt  │         only-op-equal.txt          │
                │   sec/op    │   sec/op     vs base               │
_expr-8           63.39n ± 0%   64.28n ± 0%  +1.40% (p=0.000 n=10)
_expr_reuseVm-8   44.30n ± 1%   45.11n ± 0%  +1.83% (p=0.001 n=10)
geomean           52.99n        53.85n       +1.62%
```
So if you find a ~2% improvement worthwhile, I can add `OpEqualBool`. Similarly, we can do the same with `OpLess`/`OpMore` for ints and floats.