@antonmedv 2024-03-04
> So if you find a ~2% improvement worthwhile

I think we can postpone implementing it. We can focus on features with >10% performance boost.