@antonmedv 2024-03-04
So what is the status of this PR? Will we separate array equal cases to a func `deepArrayEquals`?