@bizywizy 2024-03-04
Firstly, I'm gonna write benchmarks for `runtime.Equal` to see how the quantity of cases impacts the switch-case statement. Looks like that dividing several cases into separate function isn't really beneficial.