@bizywizy 2024-03-04
I compared master vs extended switch-case:
```
❯ benchstat master.txt switch-case.txt                                                                                                               
goos: darwin
goarch: arm64
pkg: github.com/expr-lang/expr/vm/runtime
                                         │  master.txt  │           switch-case.txt            │
                                         │    sec/op    │    sec/op     vs base                │
Equal/int_==_int-8                          2.081n ± 2%    2.119n ± 1%   +1.83% (p=0.003 n=10)
Equal/int_!=_int-8                          2.149n ± 1%    2.151n ± 0%        ~ (p=0.592 n=10)
Equal/int_==_int8-8                         2.185n ± 1%    2.201n ± 0%   +0.76% (p=0.003 n=10)
Equal/int_==_int16-8                        2.198n ± 1%    2.224n ± 1%   +1.18% (p=0.001 n=10)
Equal/int_==_int32-8                        2.179n ± 0%    2.209n ± 1%   +1.42% (p=0.000 n=10)
Equal/int_==_int64-8                        2.171n ± 1%    2.206n ± 1%   +1.59% (p=0.000 n=10)
Equal/float_==_float-8                      2.153n ± 1%    2.172n ± 1%   +0.91% (p=0.022 n=10)
Equal/float_!=_float-8                      2.162n ± 1%    2.174n ± 1%   +0.53% (p=0.027 n=10)
Equal/float_==_int-8                        2.266n ± 0%    2.317n ± 1%   +2.25% (p=0.000 n=10)
Equal/float_!=_int-8                        2.282n ± 0%    2.317n ± 0%   +1.58% (p=0.000 n=10)
Equal/string_==_string-8                    4.031n ± 0%    4.032n ± 0%        ~ (p=0.956 n=10)
Equal/string_!=_string-8                    4.308n ± 0%    4.285n ± 0%   -0.53% (p=0.002 n=10)
Equal/bool_==_bool-8                        2.253n ± 1%    2.168n ± 2%   -3.75% (p=0.000 n=10)
Equal/bool_!=_bool-8                        2.255n ± 0%    2.152n ± 1%   -4.57% (p=0.000 n=10)
Equal/[]any_==_[]int-8                      7.353n ± 1%   12.950n ± 1%  +76.13% (p=0.000 n=10)
Equal/[]any_!=_[]int-8                      7.368n ± 0%   12.920n ± 1%  +75.36% (p=0.000 n=10)
Equal/deep_[]any_==_[]any-8                103.90n ± 0%    83.84n ± 0%  -19.31% (p=0.000 n=10)
Equal/deep_[]any_!=_[]any-8                104.55n ± 1%    60.29n ± 1%  -42.34% (p=0.000 n=10)
Equal/map[string]any_==_map[string]any-8    219.1n ± 0%    214.3n ± 1%   -2.19% (p=0.000 n=10)
Equal/map[string]any_!=_map[string]any-8    56.47n ± 1%    53.26n ± 1%   -5.69% (p=0.000 n=10)
geomean                                     5.752n         5.842n        +1.57%
```
And extended switch-case vs separate function for arrays:
```
❯ benchstat switch-case.txt separate-func.txt                                                                                                        
goos: darwin
goarch: arm64
pkg: github.com/expr-lang/expr/vm/runtime
                                         │ switch-case.txt │          separate-func.txt          │
                                         │     sec/op      │   sec/op     vs base                │
Equal/int_==_int-8                             2.119n ± 1%   2.093n ± 1%        ~ (p=0.085 n=10)
Equal/int_!=_int-8                             2.151n ± 0%   2.127n ± 0%   -1.14% (p=0.000 n=10)
Equal/int_==_int8-8                            2.201n ± 0%   2.194n ± 1%        ~ (p=0.128 n=10)
Equal/int_==_int16-8                           2.224n ± 1%   2.212n ± 1%        ~ (p=0.171 n=10)
Equal/int_==_int32-8                           2.209n ± 1%   2.191n ± 1%   -0.84% (p=0.024 n=10)
Equal/int_==_int64-8                           2.206n ± 1%   2.189n ± 1%   -0.75% (p=0.008 n=10)
Equal/float_==_float-8                         2.172n ± 1%   2.172n ± 0%        ~ (p=1.000 n=10)
Equal/float_!=_float-8                         2.174n ± 1%   2.183n ± 0%   +0.44% (p=0.014 n=10)
Equal/float_==_int-8                           2.317n ± 1%   2.282n ± 1%   -1.49% (p=0.000 n=10)
Equal/float_!=_int-8                           2.317n ± 0%   2.277n ± 1%   -1.75% (p=0.000 n=10)
Equal/string_==_string-8                       4.032n ± 0%   4.024n ± 0%        ~ (p=0.324 n=10)
Equal/string_!=_string-8                       4.285n ± 0%   4.288n ± 0%        ~ (p=0.590 n=10)
Equal/bool_==_bool-8                           2.168n ± 2%   2.164n ± 1%        ~ (p=0.077 n=10)
Equal/bool_!=_bool-8                           2.152n ± 1%   2.162n ± 2%        ~ (p=0.304 n=10)
Equal/[]any_==_[]int-8                         12.95n ± 1%   16.38n ± 1%  +26.45% (p=0.000 n=10)
Equal/[]any_!=_[]int-8                         12.92n ± 1%   16.41n ± 1%  +26.97% (p=0.000 n=10)
Equal/deep_[]any_==_[]any-8                    83.84n ± 0%   98.89n ± 1%  +17.96% (p=0.000 n=10)
Equal/deep_[]any_!=_[]any-8                    60.29n ± 1%   74.37n ± 1%  +23.36% (p=0.000 n=10)
Equal/map[string]any_==_map[string]any-8       214.3n ± 1%   217.8n ± 0%   +1.66% (p=0.000 n=10)
Equal/map[string]any_!=_map[string]any-8       53.26n ± 1%   56.91n ± 1%   +6.86% (p=0.000 n=10)
geomean                                        5.842n        6.098n        +4.37%
```
The difference is just a tiny fraction of a nanosecond. I think we can say they're pretty much the same. What do you think?